requirements.txt
django-ninja
djangorestframework
psycopg2
django-simple-history
